+++ packages/runtime-web/src/api/media/audio/index.ts
@@ -1,0 +1,7 @@
+import { AudioParams, InnerAudio } from './Audio'
+
+export default {
+  createInnerAudioContext(params: AudioParams) {
+    return new InnerAudio(params)
+  }
+}